Move ManifestResults to separate file so it's easier to work with.
